v1.378 -SkillStats() de TipoSpell y TipoBDebuff ya no devuelve un array, eso desvelaria la implementacion exacta de ambos, ahora devuelve un iterator como dios manda, se añade un metodo para acceder a cada elemento llamado getSkillStat, de esta manera podemos implementar el tipo de almacenamiento de datos que queramos sin desvelarlo a las clases que acceden a el

+++ MyrranMP/MyrranCore/src/Interfaces/BDebuff/BDebuffI.java
@@ -22,7 +22,7 @@ public interface BDebuffI
     public String getNombre ();
     public String getDescripcion ();
     public boolean isDebuff ();
-    public int getStacksMaximos ();
+    public byte getStacksMaximos ();
     public TipoBDebuffI getTipoBDebuff();
     public SkillStat getSkillStat(int numSkillStat);
     public Iterator<SkillStat> getSkillStats();
